Purge events tied to closed partitions (#259)
In some cases, after a rebalance of partitions and depending on the late
arrival of acks from HEC, it's needed to remove all references to events tied
to the closed partitions. This purge has to be made in buffered events, failed
events, offsets, and topic/partition records. This is considered safe because
those events will be picked up by the task that opens the partitions closed in
the original task.